refactor: move files from pages to components
